% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons-package.R
\docType{package}
\name{ons-package}
\alias{ons}
\alias{ons-package}
\title{ons: Download Data from the 'Office for National Statistics'}
\description{
Provides functions to download and tidy statistical data published by the 'Office for National Statistics' \url{https://www.ons.gov.uk}. Covers GDP, inflation (CPI, CPIH, RPI), unemployment, employment, wages, trade, retail sales, house prices, productivity, population, and public sector finances. Most series are fetched from the 'ONS' website using its CSV time series endpoint. House price data is sourced from 'HM Land Registry' \url{https://www.gov.uk/government/organisations/land-registry}. Data is cached locally between sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/ons}
  \item Report bugs at \url{https://github.com/charlescoverdale/ons/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\concept{CPI}
\concept{GDP}
\concept{ONS}
\concept{Office for National Statistics}
\concept{UK economic data}
\concept{house prices}
\concept{inflation}
\concept{population}
\concept{trade}
\concept{unemployment}
\concept{wages}
