% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_population.R
\name{ons_population}
\alias{ons_population}
\title{Download population estimates}
\source{
\url{https://www.ons.gov.uk}
}
\usage{
ons_population(from = NULL, to = NULL, cache = TRUE)
}
\arguments{
\item{from}{Date or character (YYYY-MM-DD). Start date. Defaults to
\code{NULL} (all available data).}

\item{to}{Date or character (YYYY-MM-DD). End date. Defaults to \code{NULL}.}

\item{cache}{Logical. Use cached data if available (default \code{TRUE}).}
}
\value{
A data frame with columns:
\describe{
\item{date}{Date. Mid-year date (represented as 1 January).}
\item{value}{Numeric. Population estimate (thousands).}
}
}
\description{
Downloads mid-year population estimates for the United Kingdom from the
ONS. Returns annual estimates.
}
\examples{
\donttest{
op <- options(ons.cache_dir = tempdir())
try(ons_population(from = "2000-01-01"))
options(op)
}

}
\seealso{
Other economic indicators: 
\code{\link{ons_cpi}()},
\code{\link{ons_gdp}()},
\code{\link{ons_monthly_gdp}()},
\code{\link{ons_productivity}()}
}
\concept{economic indicators}
